# Implemented a "group by trace" processor.