--- a/processor/groupbytraceprocessor/factory.go
+++ b/processor/groupbytraceprocessor/factory.go
@@ -1,0 +1,88 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package groupbytraceprocessor
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/config/configmodels"
+	"go.opentelemetry.io/collector/consumer"
+	"go.opentelemetry.io/collector/processor/processorhelper"
+)
+
+const (
+	// typeStr is the value of "type" for this processor in the configuration.
+	typeStr configmodels.Type = "groupbytrace"
+)
+
+var (
+	defaultWaitDuration   = time.Second
+	defaultNumTraces      = 1_000_000
+	defaultDiscardOrphans = false
+	defaultStoreOnDisk    = false
+
+	errDiskStorageNotSupported    = fmt.Errorf("option 'disk storage' not supported in this release")
+	errDiscardOrphansNotSupported = fmt.Errorf("option 'discard orphans' not supported in this release")
+)
+
+// NewFactory returns a new factory for the Filter processor.
+func NewFactory() component.ProcessorFactory {
+	return processorhelper.NewFactory(
+		typeStr,
+		createDefaultConfig,
+		processorhelper.WithTraces(createTraceProcessor))
+}
+
+// createDefaultConfig creates the default configuration for the processor.
+func createDefaultConfig() configmodels.Processor {
+	return &Config{
+		ProcessorSettings: configmodels.ProcessorSettings{
+			TypeVal: typeStr,
+			NameVal: string(typeStr),
+		},
+		NumTraces:    defaultNumTraces,
+		WaitDuration: defaultWaitDuration,
+
+		// not supported for now
+		DiscardOrphans: defaultDiscardOrphans,
+		StoreOnDisk:    defaultStoreOnDisk,
+	}
+}
+
+// createTraceProcessor creates a trace processor based on this config.
+func createTraceProcessor(
+	_ context.Context,
+	params component.ProcessorCreateParams,
+	cfg configmodels.Processor,
+	nextConsumer consumer.TraceConsumer) (component.TraceProcessor, error) {
+
+	oCfg := cfg.(*Config)
+
+	var st storage
+	if oCfg.StoreOnDisk {
+		return nil, errDiskStorageNotSupported
+	}
+	if oCfg.DiscardOrphans {
+		return nil, errDiscardOrphansNotSupported
+	}
+
+	// the only supported storage for now
+	st = newMemoryStorage()
+
+	return newGroupByTraceProcessor(params.Logger, st, nextConsumer, *oCfg)
+}