--- a/service/defaultcomponents/defaults_test.go
+++ b/service/defaultcomponents/defaults_test.go
@@ -51,6 +51,7 @@ func TestDefaultComponents(t *testing.T) {
 		"probabilistic_sampler",
 		"span",
 		"filter",
+		"groupbytrace",
 	}
 	expectedExporters := []configmodels.Type{
 		"opencensus",