… processor/groupbytraceprocessor/README.md
… processor/groupbytraceprocessor/event_test.go
… processor/groupbytraceprocessor/factory_test.go
… processor/groupbytraceprocessor/processor_test.go
… processor/groupbytraceprocessor/ring_buffer_test.go
… processor/groupbytraceprocessor/storage_memory_test.go
… processor/groupbytraceprocessor/testdata/config.yaml