Signed-off-by: Juraci Paixão Kröhling <juraci@kroehling.de>

**Description:** This change adds a new 'groupbytrace' processor, as defined on #1309. This addresses the first item in the action plan.

**Link to tracking Issue:** #1309

**Testing:** extensive unit tests have been added

**Documentation:** sample configuration + readme added