@jpkrohling 2020-07-22
> Is this a common enough use case to go into core or should it go in contrib?

My idea is that once this processor gets stable enough, the tail-based sampler starts using it instead of having its own similar logic.