@jpkrohling 2020-07-23
I just realized that I left a change to `service/defaultcomponents/defaults.go` out. Will amend this PR later to include this, but the change there is trivial and doesn't block the review.