@jpkrohling 2020-07-27
> Using pdata model without conversion to OTLP might be better, and that will allow you will to use TraceID.String() instead of hashing.

The last commit changes the PR to use the internal pdata model instead of OTLP. The performance gain seems to be consistent and relevant:

Before the last commit (f81445e701d508d33c6f8be79e509f6c7683e9ac), 3 benchmark runs:
```
$ go test -benchmem -run="^$" ./processor/groupbytraceprocessor -bench "^(BenchmarkConsumeTracesCompleteOnFirstBatch)$"
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/processor/groupbytraceprocessor
BenchmarkConsumeTracesCompleteOnFirstBatch-12    	 1000000	      7714 ns/op	   12468 B/op	     159 allocs/op
PASS
ok  	go.opentelemetry.io/collector/processor/groupbytraceprocessor	7.779s
$ go test -benchmem -run="^$" ./processor/groupbytraceprocessor -bench "^(BenchmarkConsumeTracesCompleteOnFirstBatch)$"
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/processor/groupbytraceprocessor
BenchmarkConsumeTracesCompleteOnFirstBatch-12    	 1000000	      8129 ns/op	   13159 B/op	     168 allocs/op
PASS
ok  	go.opentelemetry.io/collector/processor/groupbytraceprocessor	8.193s
$ go test -benchmem -run="^$" ./processor/groupbytraceprocessor -bench "^(BenchmarkConsumeTracesCompleteOnFirstBatch)$"
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/processor/groupbytraceprocessor
BenchmarkConsumeTracesCompleteOnFirstBatch-12    	 1000000	      8736 ns/op	   13946 B/op	     178 allocs/op
PASS
ok  	go.opentelemetry.io/collector/processor/groupbytraceprocessor	8.802s
```

After:
```
$ go test -benchmem -run="^$" ./processor/groupbytraceprocessor -bench "^(BenchmarkConsumeTracesCompleteOnFirstBatch)$"
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/processor/groupbytraceprocessor
BenchmarkConsumeTracesCompleteOnFirstBatch-12    	 1000000	      3377 ns/op	    6463 B/op	      54 allocs/op
PASS
ok  	go.opentelemetry.io/collector/processor/groupbytraceprocessor	3.438s
$ go test -benchmem -run="^$" ./processor/groupbytraceprocessor -bench "^(BenchmarkConsumeTracesCompleteOnFirstBatch)$"
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/processor/groupbytraceprocessor
BenchmarkConsumeTracesCompleteOnFirstBatch-12    	 1000000	      2781 ns/op	    5020 B/op	      44 allocs/op
PASS
ok  	go.opentelemetry.io/collector/processor/groupbytraceprocessor	2.842s
$ go test -benchmem -run="^$" ./processor/groupbytraceprocessor -bench "^(BenchmarkConsumeTracesCompleteOnFirstBatch)$"
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/collector/processor/groupbytraceprocessor
BenchmarkConsumeTracesCompleteOnFirstBatch-12    	 1000000	      3560 ns/op	    7062 B/op	      59 allocs/op
PASS
ok  	go.opentelemetry.io/collector/processor/groupbytraceprocessor	3.636s
```