@jpkrohling 2020-07-29
@nilebox, @owais: other than the clarification around the shutdown behavior, is there anything pending for this PR?