@jpkrohling 2020-07-30
Thank you @nilebox for the review rounds and for your patience. I believe I addressed all issues, except for the ones that deserve further discussion, namely:

- [ ] what should happen with in-flight traces when the processor is shutting down? Should we just discard them? Should we flush them immediately, blocking until they are consumed, possibly honoring the context's deadline? 
- [ ] should the processor protect itself against nil next consumers? Other processors seem to be doing it, others don't. The fact that there is an explicit error called `componenterror.ErrNilNextConsumer` makes me wonder what the contract really is, even though it doesn't make sense to have a processor without next consumer (which should at least be an exporter)
- [ ] can `ResourceSpans/InstrumentationLibrarySpans` contain spans from multiple traces?