@nilebox 2020-07-31
The code looks good to me now, but need to check the remaining open questions with maintainers of the repo.
Also not sure if this processor should live in the core repo or contrib.