@jpkrohling 2020-07-31
>  Also not sure if this processor should live in the core repo or contrib.

This has been asked before by @jrcamp, and here's my answer:

> My idea is that once this processor gets stable enough, the tail-based sampler starts using it instead of having its own similar logic.