@jpkrohling 2020-07-31
The last commit also removed the nil check on the next consumer, as I agree that it doesn't really make sense for a processor to not have a next consumer.

The only remaining point is about the shutdown then.

@tigrannajaryan, @bogdandrutu, would you please help us on making a decision here? [Basically](https://github.com/open-telemetry/opentelemetry-collector/pull/1362#issuecomment-666402837): 

>  what should happen with in-flight traces when the processor is shutting down? Should we just discard them? Should we flush them immediately, blocking until they are consumed, possibly honoring the context's deadline?