@jpkrohling 2020-07-31
@nilebox would it be OK if I create an issue to discuss the shutdown behavior, implementing a fix in a follow-up PR?