@jpkrohling 2020-08-10
I addressed @bogdandrutu's comment and squashed all the commits, since this was probably the last change in this PR.